+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Key: [+] added, [-] removed, [~] modified, [!] important.
 
+## 202007r1
++ [+] Added [__playnext__] command which allows users to add a song/playlist to the front of the current queue.
+
 ## 202006r1
 + [~] Improve [__clear__] command usability.
 + [~] Update dependencies.